+++ apps/backend/src/modules/courses/libs/types/types.ts
@@ -1,6 +1,7 @@
 export {
 	type AddCourseRequestDto,
 	type CourseDto,
+	type CourseGetAllByUserRequestDto,
 	type CourseSearchRequestDto,
 	type CoursesResponseDto,
 } from "@trackmates/shared";