+++ apps/backend/src/modules/section-statuses/libs/types/types.ts
@@ -1,0 +1,7 @@
+export {
+	type SectionStatusAddRequestDto,
+	type SectionStatusGetAllRequestDto,
+	type SectionStatusGetAllResponseDto,
+	type SectionStatusResponseDto,
+	type SectionStatusUpdateRequestDto,
+} from "@trackmates/shared";