+++ apps/backend/src/modules/section-statuses/libs/validation-schemas/validation-schemas.ts
@@ -1,0 +1,6 @@
+export {
+	sectionStatusCreateBodyValidationSchema,
+	sectionStatusGetAllQueryValidationSchema,
+	sectionStatusUpdateBodyValidationSchema,
+	sectionStatusUpdateQueryValidationSchema,
+} from "@trackmates/shared";