+++ apps/backend/src/modules/section-statuses/section-status.model.ts
@@ -1,0 +1,37 @@
+import { Model, type RelationMappings } from "objection";
+
+import {
+	AbstractModel,
+	DatabaseTableName,
+} from "~/libs/modules/database/database.js";
+import { type ValueOf } from "~/libs/types/types.js";
+import { CourseSectionModel } from "~/modules/course-sections/course-sections.js";
+
+import { type SectionStatus } from "./libs/enums/enums.js";
+
+class SectionStatusModel extends AbstractModel {
+	public static relationMappings = (): RelationMappings => {
+		return {
+			courseSections: {
+				join: {
+					from: `${DatabaseTableName.SECTION_STATUSES}.courseSectionId`,
+					to: `${DatabaseTableName.COURSE_SECTIONS}.id`,
+				},
+				modelClass: CourseSectionModel,
+				relation: Model.BelongsToOneRelation,
+			},
+		};
+	};
+
+	public courseSectionId!: number;
+
+	public status!: ValueOf<typeof SectionStatus>;
+
+	public userId!: number;
+
+	public static override get tableName(): string {
+		return DatabaseTableName.SECTION_STATUSES;
+	}
+}
+
+export { SectionStatusModel };