+++ apps/frontend/src/libs/helpers/helpers.ts
@@ -1,8 +1,10 @@
 export { checkIfPathMatchingPattern } from "./check-if-path-matching-pattern/check-if-path-matching-pattern.helper.js";
+export { getTruncatedContent } from "./get-truncated-content/get-truncated-content.helper.js";
 export { getValidClassNames } from "./get-valid-class-names/get-valid-class-names.helper.js";
 export {
 	configureString,
 	getDifferenceInHours,
 	getFormattedDate,
+	getSanitizedHtml,
 	initDebounce,
 } from "@trackmates/shared";