+++ apps/frontend/src/pages/friends/libs/components/friend/friend.tsx
@@ -1,5 +1,5 @@
+import defaultAvatar from "~/assets/img/default-avatar.png";
 import { Button, Image, Link } from "~/libs/components/components.js";
-import { DEFAULT_USER_AVATAR } from "~/libs/constants/constants.js";
 import { type AppRoute } from "~/libs/enums/enums.js";
 import {
 	useAppDispatch,
@@ -64,7 +64,7 @@ const Friend: React.FC<Properties> = ({ friend }: Properties) => {
 				<Image
 					alt="User avatar"
 					className={styles["portrait"]}
-					src={friend.avatarUrl ?? DEFAULT_USER_AVATAR}
+					src={friend.avatarUrl ?? defaultAvatar}
 				/>
 				<p
 					className={styles["fullName"]}
@@ -78,6 +78,7 @@ const Friend: React.FC<Properties> = ({ friend }: Properties) => {
 					label={isFollowing ? "Following" : "Follow"}
 					onClick={isFollowing ? handleUnfollow : handleFollow}
 					size="small"
+					style={isFollowing ? "secondary" : "primary"}
 				/>
 				<Button
 					className={styles["start-chat"]}